ai_memory_system/requirements.txt
fastapi
uvicorn
chromadb
huggingface_hub
pydantic
python-dotenv
torch
